+++ .gitignore
@@ -1,5 +1,6 @@
 *.pyc
 .vscode
+htmlcov
 htmlcov/*
 !htmlcov/.keep
 pgdata
