+++ Dockerfile
@@ -18,10 +18,10 @@ COPY requirements.txt .
 
 RUN pip install -r requirements.txt
 
-
+COPY scripts/ scripts/
+RUN chmod +x scripts/command-dev.sh
 COPY djangoplicity/ djangoplicity/
 COPY test_project/ test_project/
-COPY test_app/ test_app/
 COPY setup.py .
 COPY tox.ini .
 COPY .coveragerc .
