+++ requirements/requirements-testing.txt
@@ -1,8 +1,3 @@
 #test dependecies
 coverage
 tox
-pytest>=4.6.1,<4.7
-pytest-django==3.9.0
-pytest-cov==2.9.0
-pytest-timeout==1.3.4
-sqlparse
