+++ scripts/command-dev.sh
@@ -1,2 +1,3 @@
-python manage.py migrate
-python manage.py runserver 0.0.0.0:8000
+#!/bin/sh
+python ./manage.py migrate
+python ./manage.py runserver 0.0.0.0:8000
